requirements.txt
pandas
numpy
matplotlib
seaborn
prophet
scikit-learn
ydata-profiling
plotly
streamlit
langchain
openai
python-dotenv
pydantic
fastapi
uvicorn
langchain
langchain-openai
langchain-experimental
openai